DB Entities update

+++ src/main/java/com/razykrashka/bot/db/repo/MeetingRepository.java
@@ -4,6 +4,6 @@ package com.razykrashka.bot.db.repo;
 import com.razykrashka.bot.db.entity.Meeting;
 import org.springframework.data.repository.CrudRepository;
 
-public interface MeetingRepository extends CrudRepository<Meeting, Long> {
+public interface MeetingRepository extends CrudRepository<Meeting, Integer> {
 
 }
